+++ src/impl/random_variable.h
@@ -50,7 +50,7 @@ double rvs_sample(RandomVariables *rvs, uint64_t k, int thread_index,
 void rvs_reset(RandomVariables *rvs);
 bool rvs_mark_as_epigon_if_similar(RandomVariables *rvs, int leader, int i);
 bool rvs_is_epigon(const RandomVariables *rvs, int i);
-int rvs_get_num_rvs(const RandomVariables *rvs);
+uint64_t rvs_get_num_rvs(const RandomVariables *rvs);
 uint64_t rvs_get_total_samples(const RandomVariables *rvs);
 
 #endif
